move get_cfg set_cfg to log_writer
