+++ tester/__init__.py
@@ -12,8 +12,8 @@ __all__ = [
     'APIConfig', 
     'analyse_configs', 
     'USE_CACHED_NUMPY',
-    'cached_numpy', 
-    'get_cfg', 
+    'cached_numpy',
+    'get_cfg',
     'set_cfg'
 ]
 
@@ -28,21 +28,10 @@ if TYPE_CHECKING:
         APIConfig,
         analyse_configs,
         USE_CACHED_NUMPY,
-        cached_numpy
+        cached_numpy,
+        get_cfg,
+        set_cfg
     )
-    
-# Command line arguments configuration
-CMD_CONFIG = None
-
-def get_cfg():
-    global CMD_CONFIG
-    return CMD_CONFIG
-
-def set_cfg(cfg):
-    global CMD_CONFIG
-    if cfg.id != "":
-        cfg.id = "_"+cfg.id
-    CMD_CONFIG = cfg
 
 def __getattr__(name: str) -> Any:
     if name not in __all__:
@@ -79,10 +68,10 @@ def __getattr__(name: str) -> Any:
         from .api_config import cached_numpy
         return cached_numpy
     elif name == 'get_cfg':
-        from .config import get_cfg
+        from .api_config import get_cfg
         return get_cfg
     elif name == 'set_cfg':
-        from .config import set_cfg
+        from .api_config import set_cfg
         return set_cfg
 
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
