+++ tester/api_config/__init__.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING, Any
 
-__all__ = ['TensorConfig', 'APIConfig', 'analyse_configs', 'USE_CACHED_NUMPY', 'cached_numpy']
+__all__ = ['TensorConfig', 'APIConfig', 'analyse_configs', 'USE_CACHED_NUMPY', 'cached_numpy', 'get_cfg', 'set_cfg']
 
 if TYPE_CHECKING:
     from .config_analyzer import (
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
         USE_CACHED_NUMPY,
         cached_numpy
     )
-
+    from .log_writer import get_cfg, set_cfg
 def __getattr__(name: str) -> Any:
     if name not in __all__:
         raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
@@ -30,5 +30,11 @@ def __getattr__(name: str) -> Any:
     elif name == 'cached_numpy':
         from .config_analyzer import cached_numpy
         return cached_numpy
+    elif name == 'get_cfg':
+        from .log_writer import get_cfg
+        return get_cfg
+    elif name == 'set_cfg':
+        from .log_writer import set_cfg
+        return set_cfg
 
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
